mesh

+++ devCFDriver/devCFDriver/iCFDriver.h
@@ -17,7 +17,7 @@ extern "C"
 	extern void RecvVof(int len, float* vof);
 	extern void SendFile(const char* filename);
 	extern void RecvFile();
-	extern void SendData(int len, float* data);	//Backup function
-	extern void RecvData(int len, float* data);	//Backup function
+	extern void SendData(int len, double* data);	//Backup function
+	extern void RecvData(int len, double* data);	//Backup function
 
 }
